streamline axes ops pipe
